WIP: get pytest importable again
